backend/requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9,<1
pillow>=10,<12
numpy>=1.26,<3
ultralytics>=8.2,<9
